requirements.txt
beautifulsoup4==4.11.1
more_itertools==8.14.0
Pillow==9.2.0
PySide6==6.3.1
python_magic==0.4.27
requests==2.28.1
sqlparse==0.4.2
